+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [UNRELEASED]
 ### Added
 - token module
+- set / getPluginConfig methods
+- token module state dumping
 
 ### Changed
 - sdk version